requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
dnspython==2.7.0
email_validator==2.2.0
idna==3.10
numpy==2.0.2
pandas==2.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
